# src/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "One place to know about company events, activities and etc.",
        "version": "1.0.0",
        "title": "Intranet Portal",
        "contact": {
            "email": "sameer_shaik@merilytics.com"
        }
    },
    "host": "localhost:3008",
    "basePath": "/",
...
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/interaction_forums/pollingExcelDownload": {
            "post": {
                "tags": [
                    "Interaction Forums"
                ],
                "summary": "Employee silo view",
                "description": "Employee silo view",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "pass the specified fields in sample data",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/interactionForumsId"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/siloView"
                            }
                        }
                    },
                    "404": {
                        "description": "Error occurred while fetching"
                    }
                }
            }
        },
        "/interaction_forums/getInteractionForumsTrendingDetails": {
            "post": {
                "tags": [
                    "Interaction Forums"
                ],
                "summary": "Interaction Forums data display",
                "description": "Interaction Forums data display",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "pass the specified fields in sample data",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/interactionForumsPagination"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/interactionForumsDataDisplay"
                            }
                        }
                    },
                    "404": {
                        "description": "Error occurred while fetching"
                    }
                }
            }
        },
        "/interaction_forums/getCategoryDropDown": {
            "get": {
                "tags": [
                    "Interaction Forums"
                ],
                "summary": "Category drop down.",
                "description": "Category drop down.",
                "produces": [
                    "application/json"
                ],
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/categoryDropDown"
                            }
                        }
                    },
                    "404": {
                        "description": "Error occurred while fetching"
                    }
                }
            }
        },
        "/interaction_forums/getInteractionForumsNewlyAddedDetails": {
            "post": {
                "tags": [
                    "Interaction Forums"
                ],
                "summary": "Interaction Forums data display",
                "description": "Interaction Forums data display",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "pass the specified fields in sample data",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/interactionForumsPagination"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/interactionForumsDataDisplay"
                            }
                        }
                    },
                    "404": {
                        "description": "Error occurred while fetching"
                    }
                }
            }
        },
        "/interaction_forums/getInteractionForumsMostCommentedDetails": {
            "post": {
                "tags": [
                    "Interaction Forums"
                ],
                "summary": "Interaction Forums data display",
                "description": "Interaction Forums data display",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "pass the specified fields in sample data",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/interactionForumsPagination"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/interactionForumsDataDisplay"
                            }
                        }
                    },
                    "404": {
                        "description": "Error occurred while fetching"
                    }
                }
            }
        },
        "/interaction_forums/getInteractionForumsDataByID": {
            "post": {
                "tags": [
                    "Interaction Forums"
                ],
                "summary": "Interaction Forums data display",
                "description": "Interaction Forums data display",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "pass the specified fields in sample data",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/interactionsForumsId"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/interactionForumsDataDisplay"
                            }
                        }
                    },
                    "404": {
                        "description": "Error occurred while fetching"
                    }
                }
            }
        },
        "/interaction_forums/uploadImageToExistingDiscussion": {
            "post": {
                "tags": [
                    "Interaction Forums"
                ],
                "summary": "Interaction Forums data display",
                "description": "Interaction Forums data display",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "pass the specified fields in sample data",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/interactionsForumsImageUpload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/message"
                            }
                        }
                    },
                    "404": {
                        "description": "Error occurred while fetching"
                    }
                }
            }
        },
        "/interaction_forums/postAnDiscussion": {
            "post": {
                "tags": [
                    "Interaction Forums"
                ],
                "summary": "Interaction Forums data display",
                "description": "Interaction Forums data display",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "pass the specified fields in sample data",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/interactionsForumsStartDiscussion"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/message"
                            }
                        }
                    },
                    "404": {
                        "description": "Error occurred while fetching"
                    }
                }
            }
        },
        "/interaction_forums/savePollingOption": {
            "post": {
                "tags": [
                    "Interaction Forums"
                ],
                "summary": "Interaction Forums data display",
                "description": "Interaction Forums data display",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "pass the specified fields in sample data",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/savePollingOption"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/message"
                            }
                        }
                    },
                    "404": {
                        "description": "Error occurred while fetching"
                    }
                }
            }
        },
        "/interaction_forums/likeOrDislike": {
            "post": {
                "tags": [
                    "Interaction Forums"
                ],
                "summary": "Interaction Forums data display",
                "description": "Interaction Forums data display",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "pass the specified fields in sample data",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/likeOrDislike"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/message"
                            }
                        }
                    },
                    "404": {
                        "description": "Error occurred while fetching"
                    }
                }
            }
        },
        "/interaction_forums/getPollingEmployeeInfo": {
            "post": {
                "tags": [
                    "Interaction Forums"
                ],
                "summary": "Interaction Forums data display",
                "description": "Interaction Forums data display",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "pass the specified fields in sample data",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pollingOptions"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/pollingInfo"
                            }
                        }
                    },
                    "404": {
                        "description": "Error occurred while fetching"
                    }
                }
            }
        },
        "/interaction_forums/startOrJoinDiscussion": {
            "post": {
                "tags": [
                    "Interaction Forums"
                ],
                "summary": "Interaction Forums data display",
                "description": "Interaction Forums data display",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "pass the specified fields in sample data",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/startOrJoinDiscussion"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/interactionForumsDataDisplay"
                            }
                        }
                    },
                    "404": {
                        "description": "Error occurred while fetching"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "petstore_auth": {
            "type": "oauth2",
            "authorizationUrl": "http://petstore.swagger.io/oauth/dialog",
            "flow": "implicit",
            "scopes": {
                "write:pets": "modify pets in your account",
                "read:pets": "read your pets"
            }
        },
        "api_key": {
            "type": "apiKey",
            "name": "api_key",
            "in": "header"
        }
    },
    "definitions": {
...
        "message": {
            "type": "object",
            "properties": {},
            "example": {
                "message": "Appreciations sent successfully."
            }
        },
...
        "siloView": {
            "type": "object",
            "properties": {},
            "example": [
                {
                    "EmployeeID": 21100,
                    "employeeName": "Paavan Choudary",
                    "immediateReportingManagerID": null,
                    "designation": "CEO",
                    "emailID": "paavan@merilytics.com",
                    "dateOfJoining": "2011-01-01T00:00:00.000Z",
                    "EmpLevel": -2
                }
            ]
        },
        "interactionForumsDataDisplay": {
            "type": "object",
            "properties": {},
            "example": [
                {
                    "result": [
                        {
                            "interactionForumsId": 4,
                            "newsAndUpdatesId": 58,
                            "categoryId": 3,
                            "categoryName": "News & Updates",
                            "eventName": "Best Employee Award",
                            "colourCode": "#98B243 ",
                            "eventDescription": "This year best employee Award goes to XYZ.Congrtas!",
                            "eventStartDate": 1533624709380,
                            "eventEndDate": 1535730729777,
                            "createdBy": 21100,
                            "createdByName": "Paavan Choudary",
                            "updatedBy": 21102,
                            "updatedByName": "Srujandfsdf Gottemukkala",
                            "createdDate": 1533570866723,
                            "updatedDate": 1533657278993,
                            "CLDData": [
                                {
                                    "commentsCount": 4,
                                    "likesCount": 1,
                                    "dislikesCount": 1
                                }
                            ],
                            "imageData": [
                                {
                                    "imageId": 43,
                                    "imageName": "Best Employee Image",
                                    "imageType": "type",
                                    "image": "best employee"
                                },
                                {
                                    "imageId": 50,
                                    "imageName": "Best Employee",
                                    "imageType": "type",
                                    "image": "best employee Image"
                                },
                                {
                                    "imageId": 53,
                                    "imageName": "best employee",
                                    "imageType": "type ",
                                    "image": "best employee"
                                }
                            ]
                        }
                    ],
                    "pages": 1,
                    "rowCount": 1
                }
            ]
        },
        "interactionForumsPagination": {
            "type": "object",
            "properties": {
                "search": {
                    "type": "string"
                },
                "categoryId": {
                    "type": "integer",
                    "format": "int64"
                },
                "pageNo": {
                    "type": "integer",
                    "format": "int64"
                },
                "pageSize": {
                    "type": "integer",
                    "format": "int64"
                }
            },
            "example": {
                "search": null,
                "categoryId": 3,
                "pageNo": 1,
                "pageSize": 10,
                "employeeId": 21258
            }
        },
        "categoryDropDown": {
            "type": "object",
            "properties": {},
            "example": {
                "categoryId": 1,
                "categoryName": "Events",
                "typeId": 1,
                "colourCode": "#EE603B ",
                "categoryURL": "events",
                "serviceURL": "news&updates/eventDetails"
            }
        },
        "interactionsForumsId": {
            "type": "object",
            "properties": {
                "interactionForumsId": {
                    "type": "integer",
                    "format": "int64"
                }
            },
            "example": {
                "interactionForumsId": 22,
                "employeeId": 21258,
                "categoryId": 7
            }
        },
        "interactionsForumsImageUpload": {
            "type": "object",
            "properties": {},
            "example": {
                "imageData": [
                    {
                        "imageType": "type",
                        "imageName": "my new image",
                        "Image": "s;test data",
                        "updatedBy": 21258,
                        "createdBy": 21258
                    },
                    {
                        "imageType": "test",
                        "imageName": "test image",
                        "Image": "s;test data",
                        "updatedBy": 21258,
                        "createdBy": 21258
                    }
                ],
                "interactionForumsId": 48
            }
        },
        "interactionsForumsStartDiscussion": {
            "type": "object",
            "properties": {},
            "example": {
                "categoryId": 1,
                "eventStartDate": "2018/08/03 16:12:14",
                "eventEndDate": "2018/08/03 16:12:14",
                "eventName": "discussion name",
                "eventDescription": "discussion",
                "updatedBy": 21258,
                "createdBy": 21258,
                "imageData": [
                    {
                        "imageType": "type",
                        "imageName": "my new image",
                        "Image": "s;test data",
                        "updatedBy": 21258,
                        "createdBy": 21258
                    },
                    {
                        "imageType": "test",
                        "imageName": "test image",
                        "Image": "s;test data",
                        "updatedBy": 21258,
                        "createdBy": 21258
                    }
                ],
                "pollingOptions": [
                    {
                        "optionDescription": "option1",
                        "updatedBy": 21258,
                        "createdBy": 21258
                    },
                    {
                        "optionDescription": "option2",
                        "updatedBy": 21258,
                        "createdBy": 21258
                    }
                ]
            }
        },
        "savePollingOption": {
            "type": "object",
            "properties": {},
            "example": {
                "optionOptedId": 16,
                "optionId": 2,
                "optedBy": 21258,
                "createdBy": 21258,
                "updatedBy": 21258
            }
        },
        "likeOrDislike": {
            "type": "object",
            "properties": {},
            "example": {
                "interactionForumsId": 22,
                "employeeId": 21258,
                "liked": true
            }
        },
        "pollingInfo": {
            "type": "object",
            "properties": {},
            "example": {
                "optionOptedId": 2,
                "employeeId": 21258,
                "employeeName": "Sameer Shaik",
                "employeeDesignation": "Analyst (IT)",
                "workLocation": "KKR Square",
                "optionId": 2,
                "optionDescription": "Yes I will attend",
                "createdDate": "2018-09-21T16:18:25.707Z"
            }
        },
        "pollingOptions": {
            "type": "object",
            "properties": {},
            "example": {
                "interactionForumsId": 47,
                "optionId": 2
            }
        },
        "startOrJoinDiscussion": {
            "type": "object",
            "properties": {},
            "example": {
                "newsAndUpdatesId": 89,
                "enteredTheDiscussion": false,
                "createdBy": 21258,
                "updatedBy": 21258,
                "employeeId": 21258
            }
        },
...
        "interactionForumsId": {
            "type": "object",
            "properties": {},
            "example": {
                "interactionForumsId": 47
            }
        }
    },
...
}